refactor(Revision): move revision tables to models
Move the RAMSTK Program database table files associated with the
Revision module (RAMSTKRevision, RAMSTKFailureDefinition, RAMSTKMission,
RAMSTKMissionPhase, and RAMSTKEnvironment) to the new models/programdb
package.
