+++ src/ramstk/dao/DAO.py
@@ -16,9 +16,10 @@ from sqlalchemy.orm import scoped_session, sessionmaker
 # RAMSTK Package Imports
 from ramstk.Exceptions import DataAccessError
 from ramstk.models.programdb import (
-    RAMSTKNSWC, RAMSTKAllocation, RAMSTKDesignElectric,
-    RAMSTKDesignMechanic, RAMSTKHardware, RAMSTKMatrix,
-    RAMSTKMilHdbkF, RAMSTKReliability, RAMSTKSimilarItem
+    RAMSTKNSWC, RAMSTKAllocation, RAMSTKDesignElectric, RAMSTKDesignMechanic,
+    RAMSTKEnvironment, RAMSTKFailureDefinition, RAMSTKHardware, RAMSTKMatrix,
+    RAMSTKMilHdbkF, RAMSTKMission, RAMSTKMissionPhase, RAMSTKReliability,
+    RAMSTKRevision, RAMSTKSimilarItem
 )
 
 # RAMSTK Local Imports
@@ -29,15 +30,13 @@ from .commondb import (
     RAMSTKStatus, RAMSTKSubCategory, RAMSTKType, RAMSTKUser
 )
 from .programdb import (
-    RAMSTKAction, RAMSTKCause, RAMSTKControl, RAMSTKEnvironment,
-    RAMSTKFailureDefinition, RAMSTKFunction, RAMSTKGrowthTest,
+    RAMSTKAction, RAMSTKCause, RAMSTKControl, RAMSTKFunction, RAMSTKGrowthTest,
     RAMSTKHazardAnalysis, RAMSTKIncident, RAMSTKIncidentAction,
-    RAMSTKIncidentDetail, RAMSTKMechanism, RAMSTKMission, RAMSTKMissionPhase,
-    RAMSTKMode, RAMSTKOpLoad, RAMSTKOpStress, RAMSTKProgramInfo,
-    RAMSTKProgramStatus, RAMSTKRequirement, RAMSTKRevision, RAMSTKSoftware,
-    RAMSTKSoftwareDevelopment, RAMSTKSoftwareReview, RAMSTKSoftwareTest,
-    RAMSTKStakeholder, RAMSTKSurvival, RAMSTKSurvivalData, RAMSTKTest,
-    RAMSTKTestMethod, RAMSTKUnits, RAMSTKValidation
+    RAMSTKIncidentDetail, RAMSTKMechanism, RAMSTKMode, RAMSTKOpLoad,
+    RAMSTKOpStress, RAMSTKProgramInfo, RAMSTKProgramStatus, RAMSTKRequirement,
+    RAMSTKSoftware, RAMSTKSoftwareDevelopment, RAMSTKSoftwareReview,
+    RAMSTKSoftwareTest, RAMSTKStakeholder, RAMSTKSurvival, RAMSTKSurvivalData,
+    RAMSTKTest, RAMSTKTestMethod, RAMSTKUnits, RAMSTKValidation
 )
 from .RAMSTKCommonDB import (
     RAMSTK_CATEGORIES, RAMSTK_CONDITIONS, RAMSTK_FAILURE_MODES,
