+++ src/ramstk/dao/programdb/__init__.py
@@ -10,8 +10,6 @@
 from .RAMSTKAction import RAMSTKAction
 from .RAMSTKCause import RAMSTKCause
 from .RAMSTKControl import RAMSTKControl
-from .RAMSTKEnvironment import RAMSTKEnvironment
-from .RAMSTKFailureDefinition import RAMSTKFailureDefinition
 from .RAMSTKFunction import RAMSTKFunction
 from .RAMSTKGrowthTest import RAMSTKGrowthTest
 from .RAMSTKHazardAnalysis import RAMSTKHazardAnalysis
@@ -19,15 +17,12 @@ from .RAMSTKIncident import RAMSTKIncident
 from .RAMSTKIncidentAction import RAMSTKIncidentAction
 from .RAMSTKIncidentDetail import RAMSTKIncidentDetail
 from .RAMSTKMechanism import RAMSTKMechanism
-from .RAMSTKMission import RAMSTKMission
-from .RAMSTKMissionPhase import RAMSTKMissionPhase
 from .RAMSTKMode import RAMSTKMode
 from .RAMSTKOpLoad import RAMSTKOpLoad
 from .RAMSTKOpStress import RAMSTKOpStress
 from .RAMSTKProgramInfo import RAMSTKProgramInfo
 from .RAMSTKProgramStatus import RAMSTKProgramStatus
 from .RAMSTKRequirement import RAMSTKRequirement
-from .RAMSTKRevision import RAMSTKRevision
 from .RAMSTKSoftware import RAMSTKSoftware
 from .RAMSTKSoftwareDevelopment import RAMSTKSoftwareDevelopment
 from .RAMSTKSoftwareReview import RAMSTKSoftwareReview
