+++ src/ramstk/models/programdb/RAMSTKMissionPhase.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 #
-#       ramstk.dao.RAMSTKMissionPhase.py is part of The RAMSTK Project
+#       ramstk.models.programdb.RAMSTKMissionPhase.py is part of The RAMSTK
+#       Project
 #
 # All rights reserved.
 # Copyright 2007 - 2017 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
@@ -23,6 +24,12 @@ class RAMSTKMissionPhase(RAMSTK_BASE):
     This table shares a One-to-Many relationship with ramstk_environment.
     """
 
+    __defaults__ = {
+        'description': b'',
+        'name': '',
+        'phase_start': 0.0,
+        'phase_end': 0.0
+    }
     __tablename__ = 'ramstk_mission_phase'
     __table_args__ = {'extend_existing': True}
 
@@ -40,15 +47,23 @@ class RAMSTKMissionPhase(RAMSTK_BASE):
         nullable=False,
     )
 
-    description = Column('fld_description', BLOB, default=b'')
-    name = Column('fld_name', String(256), default='')
-    phase_start = Column('fld_phase_start', Float, default=0.0)
-    phase_end = Column('fld_phase_end', Float, default=0.0)
+    description = Column('fld_description',
+                         BLOB,
+                         default=__defaults__['description'])
+    name = Column('fld_name', String(256), default=__defaults__['name'])
+    phase_start = Column('fld_phase_start',
+                         Float,
+                         default=__defaults__['phase_start'])
+    phase_end = Column('fld_phase_end',
+                       Float,
+                       default=__defaults__['phase_end'])
 
     # Define the relationships to other tables in the RAMSTK Program database.
     mission = relationship('RAMSTKMission', back_populates='phase')
     environment = relationship(
-        'RAMSTKEnvironment', back_populates='phase', cascade='delete',
+        'RAMSTKEnvironment',
+        back_populates='phase',
+        cascade='delete',
     )
 
     is_mission = False
@@ -75,28 +90,19 @@ class RAMSTKMissionPhase(RAMSTK_BASE):
 
     def set_attributes(self, attributes):
         """
-        Set the current values of the RAMSTKMissionPhase data model attributes.
+        Set the current values of RAMSTKMissionPhase data model attributes.
 
-        :param tuple attributes: tuple containing the values to set.
-        :return:
+        .. note:: you should pop the mission ID and phase ID entries from
+        the attributes dict before passing it to this method.
+
+        :param dict attributes: dict of values to assign to the instance
+            attributes.
+        :return: None
+        :rtype: None
+        :raise: AttributeError if passed an attribute key that doesn't exist as
+            a table field.
         """
-        _error_code = 0
-        _msg = "RAMSTK SUCCESS: Updating RAMSTKMissionPhase {0:d} attributes.". \
-            format(self.phase_id)
-
-        try:
-            self.description = none_to_default(attributes['description'], b'')
-            self.name = str(none_to_default(attributes['name'], ''))
-            self.phase_start = float(
-                none_to_default(attributes['phase_start'], 0.0),
-            )
-            self.phase_end = float(
-                none_to_default(attributes['phase_end'], 0.0),
-            )
-        except KeyError as _err:
-            _error_code = 40
-            _msg = "RAMSTK ERROR: Missing attribute {0:s} in attribute " \
-                   "dictionary passed to " \
-                   "RAMSTKMissionPhase.set_attributes().".format(str(_err))
-
-        return _error_code, _msg
+        for _key in attributes:
+            getattr(self, _key)
+            setattr(self, _key,
+                    none_to_default(attributes[_key], self.__defaults__[_key]))
