+++ src/ramstk/models/programdb/__init__.py
@@ -11,9 +11,14 @@
 from .RAMSTKAllocation import RAMSTKAllocation
 from .RAMSTKDesignElectric import RAMSTKDesignElectric
 from .RAMSTKDesignMechanic import RAMSTKDesignMechanic
+from .RAMSTKEnvironment import RAMSTKEnvironment
+from .RAMSTKFailureDefinition import RAMSTKFailureDefinition
 from .RAMSTKHardware import RAMSTKHardware
 from .RAMSTKMatrix import RAMSTKMatrix
 from .RAMSTKMilHdbkF import RAMSTKMilHdbkF
+from .RAMSTKMission import RAMSTKMission
+from .RAMSTKMissionPhase import RAMSTKMissionPhase
 from .RAMSTKNSWC import RAMSTKNSWC
 from .RAMSTKReliability import RAMSTKReliability
+from .RAMSTKRevision import RAMSTKRevision
 from .RAMSTKSimilarItem import RAMSTKSimilarItem
